refactoring FirebaseApi on client
